(ProcessBinaries): Minor change to prevent warning in windows.

+++ src/id3/int28.cpp
@@ -1,4 +1,4 @@
-// $Id: int28.cpp,v 1.11 2000/04/05 05:21:15 eldamitri Exp $
+// $Id: int28.cpp,v 1.12 2000/04/10 16:57:47 eldamitri Exp $
 
 // id3lib: a C++ library for creating and manipulating id3v1/v2 tags
 // Copyright 1999, 2000  Scott Thomas Haug
@@ -50,10 +50,10 @@ void int28::set(uchar val[sizeof(uint32)])
   }
 
   uchar bytes [4];
-  bytes[0] = ((__acValue[0] >> 3) & MASK4);
-  bytes[1] = ((__acValue[1] >> 2) & MASK5) | ((__acValue[0] & MASK3) << 5);
-  bytes[2] = ((__acValue[2] >> 1) & MASK6) | ((__acValue[1] & MASK2) << 6);
-  bytes[3] = ((__acValue[3] >> 0) & MASK7) | ((__acValue[2] & MASK1) << 7);
+  bytes[0] = (uchar) ((__acValue[0] >> 3) & MASK4);
+  bytes[1] = (uchar)(((__acValue[1] >> 2) & MASK5) | ((__acValue[0] & MASK3) << 5));
+  bytes[2] = (uchar)(((__acValue[2] >> 1) & MASK6) | ((__acValue[1] & MASK2) << 6));
+  bytes[3] = (uchar)(((__acValue[3] >> 0) & MASK7) | ((__acValue[2] & MASK1) << 7));
 
   __nValue = ParseNumber(bytes);
 }
@@ -64,7 +64,7 @@ void int28::set(uint32 val)
   for (size_t i = 0; i < sizeof(uint32); i++)
   {
     __acValue[sizeof(uint32) - 1 - i] = 
-      (uchar) ((val >> (i * 7)) & MASK7) & MASK8;
+      (uchar) (((val >> (i * 7)) & MASK7) & MASK8);
   }
 }
 
@@ -94,6 +94,9 @@ istream& operator>>(istream& in, int28& val)
 }
 
 // $Log: int28.cpp,v $
+// Revision 1.12  2000/04/10 16:57:47  eldamitri
+// (ProcessBinaries): Minor change to prevent warning in windows.
+//
 // Revision 1.11  2000/04/05 05:21:15  eldamitri
 // Updated initial comment information to reflect license, copyright
 // change.
